clean code, remove TODOs
