+++ src/trading-engine/orderbook.event.js
@@ -6,7 +6,6 @@ const ORDER_BOOK_READY_EVENT = 'ORDER_BOOK_READY_EVENT';
 const GET_AGGREGATED_STATE_EVENT = 'GET_AGGREGATED_STATE';
 const GET_ORDERBOOK_STATE_EVENT = 'GET_ORDERBOOK_STATE';
 const ORDER_BOOK_EVENT = 'ORDER_BOOK_EVENT';
-// TODO: new event to get ohlcv data
 
 const REASON_OBJECT_TYPE = {
   PLACED: 'PLACED',
